requirements.txt
django
pillow
django-autoslug